Update your_first_project.md
This is a proposal to give specific explanations right at the moment the added infos are needed, as they were not given before in the tutorial. Also i think it is useful to show the differences between the usage of CMD, Powershell and Linux/Unix-shells, because it is assumed using a command-line to use the openfl it makes no sense to tell the user to copy a-yet-unknown-thingy to a yet-not-explained-location.

This is just an example for how i think it can reduce confusion in ongoing tutorial. As it is not clear which kind of formatting is pleasing openfl, it might be useful to announce formatting-wishes to editors also.

Thanks for your interest.

PAETH CLAUDIUSRAPHAEL
S466531257 BOSS